--- src/query/planner/src/plans/alter_user.rs
@@ -1,34 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-use std::sync::Arc;
-
-use common_datavalues::DataSchema;
-use common_datavalues::DataSchemaRef;
-use common_meta_types::AuthInfo;
-use common_meta_types::UserIdentity;
-use common_meta_types::UserOption;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct AlterUserPlan {
-    pub user: UserIdentity,
-    // None means no change to make
-    pub auth_info: Option<AuthInfo>,
-    pub user_option: Option<UserOption>,
-}
-
-impl AlterUserPlan {
-    pub fn schema(&self) -> DataSchemaRef {
-        Arc::new(DataSchema::empty())
-    }
-}
