--- src/query/planner/src/plans/create_user.rs
@@ -1,35 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::sync::Arc;
-
-use common_datavalues::DataSchema;
-use common_datavalues::DataSchemaRef;
-use common_meta_types::AuthInfo;
-use common_meta_types::UserIdentity;
-use common_meta_types::UserOption;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CreateUserPlan {
-    pub user: UserIdentity,
-    pub auth_info: AuthInfo,
-    pub user_option: UserOption,
-    pub if_not_exists: bool,
-}
-
-impl CreateUserPlan {
-    pub fn schema(&self) -> DataSchemaRef {
-        Arc::new(DataSchema::empty())
-    }
-}
