+++ src/query/planner/src/plans/mod.rs
@@ -13,55 +13,25 @@
 // limitations under the License.
 
 mod alter_udf;
-mod alter_user;
 mod alter_view;
 mod call;
-mod create_role;
-mod create_stage;
 mod create_udf;
-mod create_user;
 mod create_view;
 mod delete;
-mod drop_role;
-mod drop_stage;
 mod drop_udf;
-mod drop_user;
 mod drop_view;
-mod grant_privilege;
-mod grant_role;
 mod list;
 mod projection;
-mod remove_stage;
-mod revoke_privilege;
-mod revoke_role;
-mod set_role;
 mod setting;
-mod show_grants;
-mod show_roles;
 
 pub use alter_udf::AlterUDFPlan;
-pub use alter_user::AlterUserPlan;
 pub use alter_view::AlterViewPlan;
 pub use call::CallPlan;
-pub use create_role::CreateRolePlan;
-pub use create_stage::CreateStagePlan;
 pub use create_udf::CreateUDFPlan;
-pub use create_user::CreateUserPlan;
 pub use create_view::CreateViewPlan;
 pub use delete::*;
-pub use drop_role::DropRolePlan;
-pub use drop_stage::DropStagePlan;
 pub use drop_udf::DropUDFPlan;
-pub use drop_user::DropUserPlan;
 pub use drop_view::DropViewPlan;
-pub use grant_privilege::GrantPrivilegePlan;
-pub use grant_role::GrantRolePlan;
 pub use list::ListPlan;
 pub use projection::*;
-pub use remove_stage::RemoveStagePlan;
-pub use revoke_privilege::RevokePrivilegePlan;
-pub use revoke_role::RevokeRolePlan;
-pub use set_role::SetRolePlan;
 pub use setting::*;
-pub use show_grants::ShowGrantsPlan;
-pub use show_roles::ShowRolesPlan;
