--- src/query/planner/src/plans/show_grants.rs
@@ -1,26 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-use common_datavalues::prelude::*;
-use common_meta_types::PrincipalIdentity;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct ShowGrantsPlan {
-    pub principal: Option<PrincipalIdentity>,
-}
-
-impl ShowGrantsPlan {
-    pub fn schema(&self) -> DataSchemaRef {
-        DataSchemaRefExt::create(vec![DataField::new("Grants", Vu8::to_data_type())])
-    }
-}
