+++ src/query/service/src/interpreters/interpreter_show_grants.rs
@@ -18,7 +18,7 @@ use common_datablocks::DataBlock;
 use common_datavalues::prelude::*;
 use common_exception::Result;
 use common_meta_types::PrincipalIdentity;
-use common_planner::plans::ShowGrantsPlan;
+use common_sql::plans::ShowGrantsPlan;
 use common_users::RoleCacheManager;
 use common_users::UserApiProvider;
