+++ src/query/service/src/interpreters/interpreter_user_create.rs
@@ -19,7 +19,7 @@ use common_exception::Result;
 use common_meta_types::UserGrantSet;
 use common_meta_types::UserInfo;
 use common_meta_types::UserQuota;
-use common_planner::plans::CreateUserPlan;
+use common_sql::plans::CreateUserPlan;
 use common_users::UserApiProvider;
 
 use crate::interpreters::Interpreter;
