+++ src/query/service/src/interpreters/interpreter_user_stage_create.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_types::StageType;
-use common_planner::plans::CreateStagePlan;
+use common_sql::plans::CreateStagePlan;
 use common_users::UserApiProvider;
 
 use crate::interpreters::Interpreter;
