+++ src/query/service/src/interpreters/interpreter_user_stage_drop.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_types::StageType;
-use common_planner::plans::DropStagePlan;
+use common_sql::plans::DropStagePlan;
 use common_storages_stage::StageTable;
 use common_users::UserApiProvider;
 use tracing::info;
