+++ src/query/service/src/interpreters/interpreter_user_stage_remove.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 use common_catalog::table_context::TableContext;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_planner::plans::RemoveStagePlan;
+use common_sql::plans::RemoveStagePlan;
 use common_storages_stage::StageTable;
 use regex::Regex;
