+++ src/query/sql/src/planner/binder/binder.rs
@@ -28,17 +28,17 @@ use common_exception::Result;
 use common_meta_types::UserDefinedFunction;
 use common_planner::plans::AlterUDFPlan;
 use common_planner::plans::CallPlan;
-use common_planner::plans::CreateRolePlan;
 use common_planner::plans::CreateUDFPlan;
-use common_planner::plans::DropRolePlan;
-use common_planner::plans::DropStagePlan;
 use common_planner::plans::DropUDFPlan;
-use common_planner::plans::DropUserPlan;
-use common_planner::plans::ShowGrantsPlan;
-use common_planner::plans::ShowRolesPlan;
 
+use crate::plans::CreateRolePlan;
+use crate::plans::DropRolePlan;
+use crate::plans::DropStagePlan;
+use crate::plans::DropUserPlan;
 use crate::plans::Plan;
 use crate::plans::RewriteKind;
+use crate::plans::ShowGrantsPlan;
+use crate::plans::ShowRolesPlan;
 use crate::plans::UseDatabasePlan;
 use crate::BindContext;
 use crate::ColumnBinding;
