+++ src/query/sql/src/planner/plans/plan.rs
@@ -22,30 +22,15 @@ use common_datavalues::DataSchemaRef;
 use common_datavalues::DataSchemaRefExt;
 use common_datavalues::StringType;
 use common_planner::plans::AlterUDFPlan;
-use common_planner::plans::AlterUserPlan;
 use common_planner::plans::AlterViewPlan;
 use common_planner::plans::CallPlan;
-use common_planner::plans::CreateRolePlan;
-use common_planner::plans::CreateStagePlan;
 use common_planner::plans::CreateUDFPlan;
-use common_planner::plans::CreateUserPlan;
 use common_planner::plans::CreateViewPlan;
 use common_planner::plans::DeletePlan;
-use common_planner::plans::DropRolePlan;
-use common_planner::plans::DropStagePlan;
 use common_planner::plans::DropUDFPlan;
-use common_planner::plans::DropUserPlan;
 use common_planner::plans::DropViewPlan;
-use common_planner::plans::GrantPrivilegePlan;
-use common_planner::plans::GrantRolePlan;
 use common_planner::plans::ListPlan;
-use common_planner::plans::RemoveStagePlan;
-use common_planner::plans::RevokePrivilegePlan;
-use common_planner::plans::RevokeRolePlan;
-use common_planner::plans::SetRolePlan;
 use common_planner::plans::SettingPlan;
-use common_planner::plans::ShowGrantsPlan;
-use common_planner::plans::ShowRolesPlan;
 
 use crate::optimizer::SExpr;
 use crate::plans::copy_v2::CopyPlanV2;
@@ -62,19 +47,34 @@ use crate::plans::share::ShowGrantTenantsOfSharePlan;
 use crate::plans::share::ShowObjectGrantPrivilegesPlan;
 use crate::plans::share::ShowSharesPlan;
 use crate::plans::AlterTableClusterKeyPlan;
+use crate::plans::AlterUserPlan;
 use crate::plans::CreateDatabasePlan;
+use crate::plans::CreateRolePlan;
+use crate::plans::CreateStagePlan;
 use crate::plans::CreateTablePlanV2;
+use crate::plans::CreateUserPlan;
 use crate::plans::DescribeTablePlan;
 use crate::plans::DropDatabasePlan;
+use crate::plans::DropRolePlan;
+use crate::plans::DropStagePlan;
 use crate::plans::DropTableClusterKeyPlan;
 use crate::plans::DropTablePlan;
+use crate::plans::DropUserPlan;
 use crate::plans::ExistsTablePlan;
+use crate::plans::GrantPrivilegePlan;
+use crate::plans::GrantRolePlan;
 use crate::plans::KillPlan;
 use crate::plans::OptimizeTablePlan;
+use crate::plans::RemoveStagePlan;
 use crate::plans::RenameDatabasePlan;
 use crate::plans::RenameTablePlan;
+use crate::plans::RevokePrivilegePlan;
+use crate::plans::RevokeRolePlan;
+use crate::plans::SetRolePlan;
 use crate::plans::ShowCreateDatabasePlan;
 use crate::plans::ShowCreateTablePlan;
+use crate::plans::ShowGrantsPlan;
+use crate::plans::ShowRolesPlan;
 use crate::plans::TruncateTablePlan;
 use crate::plans::UndropDatabasePlan;
 use crate::plans::UndropTablePlan;
